world map, switch scene

+++ MarioBros3/PlayScene.cpp
@@ -93,7 +93,7 @@ void CPlayScene::_ParseSection_ANIMATIONS(string line)
 	for (int i = 1; i < tokens.size(); i += 2)	// why i+=2 ?  sprite_id | frame_time  
 	{
 		int sprite_id = atoi(tokens[i].c_str());
-		int frame_time = atoi(tokens[i+1].c_str());
+		int frame_time = atoi(tokens[static_cast<int64_t>(i)+1].c_str());
 		ani->Add(sprite_id, frame_time);
 	}
